Changes made: - Tests from HW3 are refactored to step design pattern
