+++ .gitignore
@@ -4,5 +4,6 @@
 /allure-results/
 /screenshots/
 /target/
+evgenii-liashenko.iml
 homework.iml
 **/.DS_Store
