+++ src/test/java/ru/training/at/hw4/pageobjects/_void/HomePage.java
@@ -49,6 +49,7 @@ public class HomePage extends BasePage {
         webDriver.switchTo().defaultContent();
     }
 
+    @Override
     public void openPage(){
         webDriver.get(url);
     }
