+++ src/test/java/ru/training/at/hw4/tests/Exercise2DifferentElementsPageTest.java
@@ -3,126 +3,59 @@ package ru.training.at.hw4.tests;
 import io.qameta.allure.Feature;
 import io.qameta.allure.Step;
 import io.qameta.allure.Story;
-import org.openqa.selenium.WebElement;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
-import ru.training.at.hw4.pageobjects._void.DifferentElementsPage;
+import ru.training.at.hw4.tests.steps.CommonTestSteps;
+import ru.training.at.hw4.tests.steps.DifferentElementsPageTestSteps;
 
 
-import java.util.ArrayDeque;
-import java.util.Deque;
-
-import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertTrue;
-
-@Feature("Different Elements page functionality")
-@Story("Different Elements page testing")
+@Feature("Different Elements page (feature)")
+@Story("Different Elements page functionality testing (story)")
 public class Exercise2DifferentElementsPageTest extends BaseTest {
 
-    DifferentElementsPage differentElementsPage;
-
+    CommonTestSteps commonTestSteps;
+    DifferentElementsPageTestSteps differentElementsPageTestSteps;
     @BeforeClass
-    @Step("Creating DifferentElementsPage object")
-    public void createDifferentElementsPageObject() {
-        differentElementsPage = new DifferentElementsPage(webDriver);
+    @Step("Creating step objects")
+    public void createStepObjects() {
+        commonTestSteps = new CommonTestSteps(webDriver);
+        differentElementsPageTestSteps = new DifferentElementsPageTestSteps(webDriver);
     }
 
     @Test(dataProvider = "credentials", dataProviderClass = UserData.class)
     public void executeTestSteps(String userName, String password) {
         //1. Open test site by URL
-        openUrl();
+        commonTestSteps.openUrl();
 
         //2. Assert Browser title
-        checkTabTitle();
+        commonTestSteps.checkTabTitle();
 
         //3. Perform login
-        signIn(userName, password);
+        commonTestSteps.signIn(userName, password);
 
         //4. Assert User name in the left-top side of screen that user is loggined
-        checkUsernameAfterLogin();
+        commonTestSteps.checkUsernameAfterLogin();
 
         //5. Open through the header menu Service -> Different Elements Page
-        openDifferentElementsPage();
+        differentElementsPageTestSteps.openDifferentElementsPage();
 
         //6. Select checkboxes
-        selectCheckboxes();
+        differentElementsPageTestSteps.selectCheckboxes();
 
         //7. Select radio
-        selectRadioButton();
+        differentElementsPageTestSteps.selectRadioButton();
 
         //8. Select in dropdown
-        selectOptionInDropdownList();
+        differentElementsPageTestSteps.selectOptionInDropdownList();
 
         //9. Assert that
         //•	for each checkbox there is an individual log row and value is corresponded to the status of checkbox
         //•	for radio button there is a log row and value is corresponded to the status of radio button
         //•	for dropdown there is a log row and value is corresponded to the selected value.
-        checkLogs();
+        differentElementsPageTestSteps.checkLogs();
 
         //10. Close Browser
         //Closing the browser is implemented in the DriverConfigurator class
     }
-
-
-    @Step("Opening test website")
-    public void openUrl() {
-        //Opening the URL with the webdriver
-        webDriver.get("https://jdi-testing.github.io/jdi-light/index.html");
-    }
-
-    @Step("Verifying tab title")
-    public void checkTabTitle() {
-        //Asserting that the tab title matches the expected
-        assertEquals(differentElementsPage.getTabTitle(), "Home Page");
-    }
-
-    @Step("Signing in user {userName} with password {password}")
-    public void signIn(String userName, String password) {
-        //Performing a sign-in
-        differentElementsPage.userAccountMenu.authorizeUser(userName, password);
-    }
-
-    @Step("Verifying displayed username")
-    public void checkUsernameAfterLogin() {
-        //Asserting that the username is displayed
-        assertTrue(differentElementsPage.userAccountMenu.displayedName.isDisplayed());
-
-        //Asserting that the displayed username matches the expected username
-        assertEquals(differentElementsPage.userAccountMenu.displayedName.getText(), "ROMAN IOVLEV");
-    }
-
-    @Step("Opening Different Elements page")
-    public void openDifferentElementsPage() {
-        differentElementsPage.openPage();
-    }
-
-    @Step("Selecting checkboxes")
-    public void selectCheckboxes() {
-        //Activating the checkboxes
-        differentElementsPage.waterCheckbox.click();
-        differentElementsPage.windCheckbox.click();
-    }
-
-    @Step("Selecting radiobutton")
-    public void selectRadioButton() {
-        //Selecting the Selen radio button
-        differentElementsPage.selenRadioButton.click();
-    }
-
-    @Step("Selecting from dropdown list")
-    public void selectOptionInDropdownList() {
-        //Setting the Yellow option in the selector
-        differentElementsPage.yellowSelectorOption.click();
-    }
-
-    @Step("Verifying displayed logs")
-    public void checkLogs() {
-        Deque<WebElement> testLog = new ArrayDeque<>(differentElementsPage.logArea.logEntries);
-        assertTrue(testLog.removeLast().getText().contains("Water: condition changed to true"));
-        assertTrue(testLog.removeLast().getText().contains("Wind: condition changed to true"));
-        assertTrue(testLog.removeLast().getText().contains("metal: value changed to Selen"));
-        assertTrue(testLog.removeLast().getText().contains("Colors: value changed to Yellow"));
-    }
-
 
 }
